テーブル定義修正

+++ db/migrate/20190304113741_create_women.rb
@@ -4,8 +4,8 @@ class CreateWomen < ActiveRecord::Migration[5.2]
       t.integer :user_id
       t.string :last_name
       t.string :first_name
-      t.float :height
-      t.float :weight
+      t.integer :height
+      t.integer :weight
       t.integer :blood_type
       t.integer :prefecture_id
       t.string :address
